backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
lxml==6.0.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-magic==0.4.27
python-pptx==1.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
xlsxwriter==3.2.5
